+++ sdk/identity/azure-identity/azure/identity/__init__.py
@@ -2,10 +2,9 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
-from ._constants import EnvironmentVariables, KnownAuthorities
+from ._constants import KnownAuthorities
 from ._credentials import (
     AuthorizationCodeCredential,
-
     CertificateCredential,
     ChainedTokenCredential,
     ClientSecretCredential,
@@ -27,7 +26,6 @@ __all__ = [
     "DefaultAzureCredential",
     "DeviceCodeCredential",
     "EnvironmentCredential",
-    "EnvironmentVariables",
     "InteractiveBrowserCredential",
     "KnownAuthorities",
     "ManagedIdentityCredential",