+++ sdk/identity/azure-identity/azure/identity/_credentials/browser.py
@@ -33,6 +33,9 @@ class InteractiveBrowserCredential(PublicClientCredential):
     :param str client_id: the application's client ID
 
     Keyword arguments
+        - *authority*: Authority of an Azure Active Directory endpoint, for example 'login.microsoftonline.com', the
+            authority for Azure Public Cloud (which is the default). :class:`~azure.identity.KnownAuthorities` defines
+            authorities for other clouds.
         - *tenant (str)*: a tenant ID or a domain associated with a tenant. Defaults to the 'organizations' tenant,
           which can authenticate work or school accounts.
         - *timeout (int)*: seconds to wait for the user to complete authentication. Defaults to 300 (5 minutes).
@@ -40,7 +43,7 @@ class InteractiveBrowserCredential(PublicClientCredential):
     """
 
     def __init__(self, client_id, **kwargs):
-        # type: (str, Any) -> None
+        # type: (str, **Any) -> None
         self._timeout = kwargs.pop("timeout", 300)
         self._server_class = kwargs.pop("server_class", AuthCodeRedirectServer)  # facilitate mocking
         super(InteractiveBrowserCredential, self).__init__(client_id=client_id, **kwargs)